changes for nvim 0.11
