+++ doc/navigator.txt
@@ -331,21 +331,6 @@ Nondefault configuration example:
           tagfile = 'tags'
           options = '-R --exclude=.git --exclude=node_modules --exclude=test --exclude=vendor --excmd=number'
         }
-        gopls = {   -- gopls setting
-          on_attach = function(client, bufnr)  -- on_attach for gopls
-            -- your special on attach here
-            -- e.g. disable gopls format because a known issue https://github.com/golang/go/issues/45732
-            print("i am a hook, I will disable document format")
-            client.resolved_capabilities.document_formatting = false
-          end,
-          settings = {
-            gopls = {gofumpt = false} -- disable gofumpt etc,
-          }
-        },
-        lua_ls = {
-          sumneko_root_path = vim.fn.expand("$HOME") .. "/github/sumneko/lua-language-server",
-          sumneko_binary = vim.fn.expand("$HOME") .. "/github/sumneko/lua-language-server/bin/macOS/lua-language-server",
-        },
         servers = {'cmake', 'ltex'}, -- by default empty, and it should load all LSP clients avalible based on filetype
         -- but if you whant navigator load  e.g. `cmake` and `ltex` for you , you
         -- can put them in the `servers` list and navigator will auto load them.
