+++ lsp/jdtls.lua
@@ -1,0 +1,8 @@
+return {
+  settings = {
+    java = {
+      signatureHelp = { enabled = true },
+      contentProvider = { preferred = 'fernflower' },
+    },
+  },
+}
