+++ lsp/omnisharp.lua
@@ -1,0 +1,3 @@
+return {
+  cmd = { 'omnisharp', '--languageserver', '--hostPID', tostring(vim.fn.getpid()) },
+}
