+++ lsp/ruff.lua
@@ -1,0 +1,6 @@
+return {
+  filetypes = { 'python' },
+  init_options = {
+    settings = { logLevel = 'info' },
+  },
+}
